jbig85 encoder/decoder performance optimizations
Optimizations submitted 2016-04-26 by Larry Bank with these comments:

“I believe that the encoding part is still compatible with your
original code but the decoding part assumes a complete buffer. Maybe
we could add a "fast decode" version for when the dev knows that have
a complete buffer to work with.

The decoder has gotten a little more attention than the encoder, but
they are at a similar point. They both perform consistently about 2x
faster than the original code. I've optimized the code in general
(~40% speedup) and added a specific optimization for when the pixel
context will remain constant for a few pixels and the accumulator has
enough in it to not require renormalization. This is the key to
doubling the speed. Images with lots of blank areas will go even
faster. It will optimize both areas of 0's and 1's. The encoder
currently optimizes only areas of 0's, but I'll finish it shortly. I
haven't finished commenting and cleaning up the code yet, but it's not
in too bad shape. It's also likely that it will fail for certain test
images; this can be taken care of in short order.

Here's the main list of my changes:

1) Inserted the arith_decode() and arith_encode() functions into the
main decode/encode loops.

2) "Flattened" access to structure pointers to allow the compiler to
use register vars instead of indirect pointer accesses. BTW - the
register keyword is deprecated, so I removed it.

3) Reduced activity in the inner loop (fewer variables are updated)

4) Reduced the context variables from 3 to 2. This allows quicker
context calculations and easier verification of the optimization
opportunity.

5) Removed extra logical operations that could be done outside the
main loop or in a more efficient manner.

6) Mirrored the lookup tables so that the index doesn't have to be
masked. They're already small, so cache/memory is not worse for it.”

[This commit differs from the file supplied by Larry Bank on
2016-04-26 only in that I have adjusted whitespace, indentation and
brace style to match the rest of the code. Item 6) refers to the
previous commit.]